Revise File types
