+++ Tests/SwiftPackageListCoreTests/WorkspaceStateTests.swift
@@ -12,9 +12,9 @@ final class WorkspaceStateTests: XCTestCase {
     func testVersion6() throws {
         let url = Bundle.module.url(forResource: "workspace-state_v6", withExtension: "json", subdirectory: "Resources")
         let unwrappedURL = try XCTUnwrap(url)
-        let workspaceState = try WorkspaceState(at: unwrappedURL)
+        let workspaceState = try WorkspaceState(url: unwrappedURL)
         
-        switch workspaceState {
+        switch workspaceState.storage {
         case .v6(let workspaceState_V6):
             XCTAssertEqual(workspaceState_V6.version, 6)
             XCTAssertEqual(workspaceState_V6.object.artifacts[0].packageRef.identity, "intercom-ios-sp")
@@ -30,7 +30,7 @@ final class WorkspaceStateTests: XCTestCase {
         let url = Bundle.module.url(forResource: "workspace-state_v999", withExtension: "json", subdirectory: "Resources")
         let unwrappedURL = try XCTUnwrap(url)
         
-        XCTAssertThrowsError(try WorkspaceState(at: unwrappedURL)) { error in
+        XCTAssertThrowsError(try WorkspaceState(url: unwrappedURL)) { error in
             XCTAssertTrue(error is RuntimeError)
             XCTAssertEqual((error as? RuntimeError)?.description, "Version 999 of workspace-state.json is not supported")
         }
